requirements.txt
langchain-core
langgraph>0.2.27
langchain
openai
requests
python-dotenv
langchain_community
langchain-ollama
streamlit